implement handle_case
